relocate method
