+++ src/app/lib/mapper.ts
@@ -71,25 +71,6 @@ const mapConnections = (journeys: Journey[], connections: Connection[], type: "d
      */
 }
 
-const mapStops = (rawData: any): Stop[] => {
-    if (!Array.isArray(rawData)) rawData = [rawData];
-
-    return rawData.map((rawStop: any) => {
-        const {evaNumber, name, canceled, additional, separation, nameParts} = rawStop;
-        return {
-            id: evaNumber,
-            name: name,
-            cancelled: canceled,
-            additional: additional || false,
-            separation: separation || false,
-            nameParts: nameParts ? nameParts.map((part: any) => ({
-                type: part.type,
-                value: part.value
-            })) : [{type: "default", value: name}]
-        };
-    });
-}
-
 const sort = (journeys: Journey[], type: "departures" | "arrivals"): Journey[] => {
     return journeys.sort((a, b) => {
         const earliest = (journey: Journey): string => {
@@ -107,4 +88,4 @@ const sort = (journeys: Journey[], type: "departures" | "arrivals"): Journey[] =
     })
 }
 
-export {mapConnections, mapStops, sort};
+export {mapConnections, sort};
